+++ conformance_tests/core/test_ipc/src/test_ipc_put_handle_helper.cpp
@@ -21,7 +21,7 @@
 namespace bipc = boost::interprocess;
 
 #ifdef __linux__
-static void child_put_device_test(int size, ze_ipc_memory_flags_t flags,
+static void child_put_device_test(size_t size, ze_ipc_memory_flags_t flags,
                                   bool is_immediate) {
   auto driver = lzt::get_default_driver();
   auto context = lzt::create_context(driver);
@@ -59,7 +59,7 @@ static void child_put_device_test(int size, ze_ipc_memory_flags_t flags,
   }
 }
 
-static void child_put_subdevice_test(int size, ze_ipc_memory_flags_t flags,
+static void child_put_subdevice_test(size_t size, ze_ipc_memory_flags_t flags,
                                      bool is_immediate) {
   auto driver = lzt::get_default_driver();
   auto context = lzt::create_context(driver);
